success api shipping more for my shop

+++ src/main/java/com/haiemdavang/AnrealShop/dto/shipping/InfoShipment.java
@@ -13,5 +13,5 @@ import lombok.NoArgsConstructor;
 public class InfoShipment {
     public AddressDto from;
     public AddressDto to;
-    public int weight;
+    public long weight;
 }
